requirements.txt
Keras==2.3.1
numpy==1.15.4
ffmpy==0.2.2
pandas==1.0.3
matplotlib==3.2.1
Pillow==7.1.2
